API calls Upcoming movies and top rated movies
API calls up until trending, top rated and upcoming movies

+++ flutter_application_1/lib/Models/ListMovies.dart
@@ -27,9 +27,9 @@ class ListMovies{
       original_title: json["original_title"], 
       overview: json["overview"], 
       poster_path: json["poster_path"], 
-      popularity: json["popularity"], 
+      popularity: json["popularity"].toDouble(), 
       release_date: json["release_date"], 
-      vote_average: json["vote_average"],
+      vote_average: json["vote_average"].toDouble(),
     );
   }
